Reorganize code around Split, AnalysisCache, and EntryAnalysis
* Move open/closed entry organization into Analyzer
* Move Split generation out of DataProcessor into Split directly
* Move per-entry analysis into EntryAnalysis
* Add AnalysisCache to hold date-based lookups (and manage string keys better)
* Keep Analyzer as the final operation and result manager
* Remove DataProcessor
